requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
executing==2.0.1
idna==3.7
inflection==0.5.1
ipython==8.25.0
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
postgres==4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==70.1.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
stripe==10.0.0
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
